requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-environ==0.10.0
whitenoise==6.7.0
requests==2.32.3
python-dotenv==1.0.1
gTTS==2.5.4
httpx==0.27.2
deep-translator==1.11.4
gunicorn==20.1.0  